test: execute authorization type checks
