+++ apps/web/tsconfig.typecheck.json
@@ -11,6 +11,7 @@
     "**/__tests__/**"
   ],
   "extends": "./tsconfig.json",
+  "files": ["lib/authorization/contract.typecheck.test.ts"],
   "include": [
     "next-env.d.ts",
     "**/*.d.ts",
